añadidos endpoints de crear un usuario, eliminar un usuario y listar un usuario en concreto

+++ src/main/java/com/lolabotona/restapi/repository/UserRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lolabotona.restapi.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
+public interface UserRepository extends JpaRepository<User, Long>{
 
 	Optional<User> findByUsername(String username);
 	
 	Boolean existsByUsername(String username);
 	
-	Boolean existsByName(String name);
+	Boolean existsByName(String name);	
 
 }
